Add `local` field to `ExecutionLocation`
Before this commit, the way to distinguish between local and remote
`ExecutionLocation` objects was depending on the location's `name`,
which was forced to be `__LOCAL__` for all local `ExecutionLocation`
objects. This method was fragile and not portable to custom `Connector`
plugins.

This commit introduces a new `local` field to the `ExecutionLocation`
object, which defaults to `False` and is set to `True` in the
`LocalConnector` logic. Plus, this field is now used across the
StreamFlow codebase to identify local locations.