Improve logging of copy methods
This commit extracts the log messages of the `BaseConnector` copy
methods into separate functions, which can be easily invoked from other
`Connector` classes. Tha aim is to uniform the logging messages for all
the StreamFLow `Connector` classes.

In addition, this commit fixes an issue with remote-to-remote copy
inside the `ContainerConnector` logic.