Fixed `Hardware` operation methods (#616)
This commit fixes the comparison operation of the `Hardware` class
and adds a unit test. A new `_check_storages` function is introduced
to check whether it is possible to compare the storages of two `Hardware`
instances. Indeed, the "bigger" `Hardware` should contain a storage for
each `mount_point` present in the "smaller" `Hardware` instance.
Otherwise, the comparison is not valid.