Optimize remote commands through `StreamFlowPath`
This commit relies on the new `StreamFlowPath` abstraction to redirect
file-based commands to the lowest possible `ExecutionLocation` in the
wrapping hierarchy, in order to meet a `local` location whenever
possible. The main benefit of this strategy is that `local` locations
support Python-based commands, which are way faster than shell-based
remote processes.