Fixed remote ports (#643)
This commit fixes the injection of remote inputs using the `cwl_utils`
objects. Also, the unit test for input injection has been improved. Now,
other than checking the path of the output token of the `InputInjectorStep,`
the input file is copied to the same location, and a checksum check is
performed.

This commit also updates the Docker MacOS  GitHub Action to the latest
version.