Fixed `ChannelOpenError` handling on the `SSHConnector` (#664)
This commit fixes an issue in the `SSHConnector` where an infinite loop
occurred whenever `asyncssh` threw a `ChannelOpenError`. The
`asyncssh.SSHConnection` object can throw this exception from the
`create_process` method in different scenarios. 

Before this commit, the assumption was that the SSH connection was always
open after this error, and it was possible to create a new channel just by
retrying on the same connection.  For this reason, the context wasn’t reset
when this exception occurred.  However, the same exception can be thrown
when the connection is lost and closed. The connection status was not checked,
and trying to create new channels on the closed connection threw the same
exception at each attempt.
Moreover, these attempts were not counted in the `SSHConnector` retry
mechanism because the context wasn't reset.

Now, when the exception is thrown and the connection is closed, the context is
reset, and a new connection is created in the next attempt. Otherwise, the
connection is still open and, in the next attempt, is used to create a new channel.

Futhermore, the `parse_hostname` and `get_param_from_file` became public
functions, allowing external plugins to rely on them when extending the
`SSHConnector` class.