Handle single-element output lists in CWL (#672)
The `cwltool` reference implementation automatically converts
single-element lists to single elements when an output port expects a
`File` or `Directory` output (even as part of a more complex union
type). This commit implements this behaviour in StreamFlow, uniforming
it to the reference implementation.

Note that this is an undocumented behaviour and may result from a
`cwltool` bug. Therefore, things could change in the future.