Fix CWL `pickValue` with `all_non_null` option
This commit fixes a wrong behaviour when a CWL step input configured
with `pickValue` of type `all_non_null` had a single input source. In
this case, since the `all_non_null` option works with lists, the input
value should not be trsnsformed from a single-element list to a single
element.