This commit fixes the `is_available` method of the `FileToken` class. The `FileToken` can have multiple paths. The default implementation assumes that if one path does not exist, the `FileToken` is not available. (If a subclass needs different behavior, it should override the method.)
If a path of the `FileToken` is not registered or is invalid in the `DataManager`, the method will return `false`. Otherwise, the existence of the path will be checked at its location. If the path does not exist, it will be invalidated in the `DataManager`, and the method will return `false`.