Improved `is_available` method of the `FileToken` class  (#696)
This commit improves the `is_available` method of the `FileToken` class.
Before this commit, no checks were made regarding the file's existence
on the location.

If a path of the `FileToken` is not registered or is invalid in the
`DataManager`, the method will return `False`. Otherwise, the existence
of the path will be checked at its location. If the path does not exist, it
will be invalidated in the `DataManager`, and the method will
return `False`.

Note: The `FileToken` can have multiple paths. The default implementation
assumes that if one path does not exist, the `FileToken` is not available
(if a subclass needs different behavior, it should override the method).