Added `ROLLBACK` status  (#697)
This commit adds the `ROLLBACK` status and improves the
`DefaultScheduler` to manage this new status. This feature is helpful
for introducing recovery mechanisms when a job fails.

The `DefaultScheduler` gives higher priority to scheduling older jobs of
the same step when they are in `ROLLBACK`. Moreover, when a job notifies
the `ROLLBACK` status, its `JobAllocation` instance is reset.