Implemented fault tolerance based on `Database` module (#169)
This commit changes the current StreamFlow fault tolerance module.
When a step fails, the new fault tolerance module builds a sub-workflow
composed of steps from the original workflow, and these steps are added
to the sub-workflow if their output tokens are lost. This method is inspired
to the "rescueDAG" methodology implemented  by DagMAN, but it adapts
the original algorithm to a hybrid execution scenario. The steps involved
and their dependencies are retrieved  from the StreamFlow `Database`,
enabling future partial workflow re-executions.