This commit fixes a race condition in local data transfers.
The same problem was resolved in #328. This issue reappeared after the introduction of the `StreamFlowPath` class. When a file transfer is needed, before it is necessary to wait for the source data location to become available. With the introduction of the `StreamFlowPath` class, this wait was moved inside the `resolve` method. However, only the `RemoteStreamFlowPath` was performing the wait. Now, the `resolve` method of the `LocalStreamFlowPath` also performs the wait before resolving the path.