Check if bindings exist in the workflow (#708)
This commit adds a check to warn the user if there are bindings
(steps or ports) in the StreamFlow config file that do not have a
corresponding entity in the workflow. This commit also refactors the
`Command` classes in the `streamflow.core` to remove circular
dependencies. Finally, this commit improves the check for circular
deployment wrappers introduced in #705 by keeping track of the
already visited deployments.