Removed comparison operators from `Hardware` class (#711)
This commit prevents the overfilling of `Hardware` resources by
fixing an erroneous hardware comparison. In addition, this commit
removes the overloaded comparison operators in the `Hardware`
class, as they could lead to ambiguous results. In lieu of this, this
commit introduces a `satisfies` method to check whether an `Hardware`
instance has enough resources to satisfy a requirement.

This commit also improves `Hardware` tests. The
`test_single_env_few_resources` unit test of the `test_scheduler.py` file
has been improved to have a more realistic case (i.e., the location does
not have exactly the needed resources). Additionally, a `finally` block has
been added to free the resources in case of a failed test. Finally, an assertion
is made to check the job status when attempting to be scheduled to ensure
that no errors are raised. The `asyncio.wait` did not raise the exception until
the task result was checked.