Fixed preparation of workdir for list and object inputs (#732)
This PR handles the adjustment of input paths after moving the data
to the job workdir. Before this PR, only files were handled, but inputs
can also be lists or objects containing files. Unit tests have been added
to check the new behaviour.

This commit also adds a check to detect the correct creation of the
job's initial directories.