Imoproved token processing logic in `CWLObjectTokenProcessor` (#758)
This commit simplifies the token processing logic in
`CWLObjectTokenProcessor`. In StreamFlow, the value of an
`ObjectToken` is a dictionary with `str` keys and `Token`
values. When executing a CWL workflow, each key in the dictionary
is associated with a different `CWLObjectTokenProcessor`.

Before this commit, the value of the token was encapsulated inside an
`ObjectToken` and passed to the inner processor. Now, token values are
propagated directly, streamlining the process and improving its
correctness.