Fixed subworkflows case in the `RunCrateProvenanceManager` (#774)
This commit addresses an issue with subworkflows in the
`RunCrateProvenanceManager`. Before this commit, the
subworkflow did not have the `CreateAction` entity, but it was
still being used in property evaluations to update a
(non-existent) action. Now, the subworkflow is skipped from
the action update.

Additionally, this commit resolves an issue in the
`CWLRunCrateProvenanceManager` related to in-place step
and workflow definitions.