Fixed `DefaultRetagTransformer` within scatter (#742)
This commit addresses an issue with the token tag of default
input values when the step is inside a scatter. Additionally, the
logic for the optional type has been integrated with the handling
of default values.

The `DefaultRetagTransformer` has three types of input ports. The
`primary` port receives the value from the previous step, if there
is one.  If there is no previous step, the port contains a token with
 `None` value and a `TerminationToken`.  The `default` port holds the
default value token, which is used when the `primary` port has a token
with `None` value or a `TerminationToken`. Finally, the sibling ports manage
the lifecycle of the step, including propagating the next token with the
appropriate tag and handling step termination when a termination token is
received. The siblings ports are the other input ports of the step (or
workflow) , excluding the other default ports and the inputs with the
`valueFrom`.

A special case is when there are no sibling ports; in this case, the primary port
manages the step life-cycle.