Refactor `TokenProcessor` and `CommandOutputProcessor` classes
This commit normalizes the behaviour of the `TokenProcessor` and
`CommandOutputProcessor` classes in terms of type handling. In addition,
it further relazes the `CWLCommandOutput` behaviour to resolve the value
of the `Future[CommandOutput]` object only when strictly necessary.

In addition, this commit removes the unnecessary
`CWLMapCommandOutputProcessor` class. The `CWLCommandOutputProcessor`
alone is sufficient to treat both single and list outputs, making the
former class an useless complication within the codebase. However, the
`MapCommandOutputProcessor` base class has been preserved to be used
within workflow languages other than CWL, if necessary.