Add support for CWL v1.3 (#802)
This commit adds preliminary support for CWL v1.3 and enables the CI/CD
to run the CWL v1.3 conformance test suite. Note that support for CWL
v1.3 is still experimental and the standard itself has not been released,
yet. Therefore, users are discouraged to implement CWL v1.3 workflows in
production at the moment.