Added a check for the chosen deployment in the tests (#814)
This commit improves the test for local development purposes,
specifically when not all deployments are used (e.g., by specifying
a subset of deployments with the `--deploys` flag).
Before this commit, some tests would fail if certain deployments
were not selected, as they required the deployment to run. Now,
the tests are skipped for missing deployments. However,
in the CI, all deployments are tested.