Fixed `recoverable` attribute in the `ListToken` and `ObjectToken` classes (#752)
This commit fixes the `recoverable` attribute, introduced in #699,
in the `ListToken` and `ObjectToken` subclasses. The value of
`recoverable` is now computed based on the `recoverable` values
of the inner tokens.

Additionally, tests have been added to specifically check the
`recoverable` attribute, including cases for lists and objects in the
`test_recover` suite.

Finally, this commit removed leftover code in the `_load` method
of the `Token` class from changes introduced in #777