Refactor handling of `File` and `Directory` literals (#811)
This commit refactors the handling of literal files and directories, moving
its logic to a self-contained function. In addition:
- It adds the check for well-defined literal files during the `InitialWorkDir`
preparation phase by calling the newly defined function.
- It fixes a problem with the `secondaryFiles` of literal files. Before this
commit, they were not created.
- It improves the unit test on remote ports, checking literal files and
directories. Before this commit, the test was only on a directory, and was
not compliant with the definition of literal directory in CWL.
- Finally, unrelated to the literal files, it adds a fix to retrieve the step
name from the job name.