Fix `cwl.output.json` behaviour
This commit improves how StreamFlow handles the `cwl.output.json`
output. Before this commit, the `Workflow` object was modified according
to the content of `cwl.output.json`, but this solution was not portable
in case of scatter steps with different `cwl.output.json` schemas.

This commit just injects a `None` token when an item is not present in
the `cwl.output.json` dictionary. Additionally, it correctly skips the
`outputBinding` section whenever an output directly comes from a
`cwl.output.json` file.