Fixed `loadContents` in the collector step (#846)
This commit resolves an issue that occurs when a `CommandLineTool`
is executed directly and its output includes `loadContents`. Before this fix,
the `TokenProcessor` in the collector step incorrectly attempted to
load file contents. However, this loading should only be handled by the
`CommandOutputProcessor`. The redundant computation can lead to
errors, particularly when the content exceeds 64KB, as specified by the
CWL standard.