Fixed `eval` method of the `DependencyResolver`  (#867)
This PR fixes an issue where the `deps` attribute value was incorrectly
overwritten in the `eval` method of the `DependencyResolver` class. This
method is called when an expression contains JavaScript code.

This issue is closely related to the fix in issue #839, which addresses the
same problem in a non-JavaScript context (e.g., `$(inputs.a)-$(inputs.b)`).