Handle location storage retrieval errors (#871)
This commit fixes the retrieval of location storage information.
The code now handles output parsing failures, such as when an error
message is returned because the user does not have permission to
access information about a partition. Additionally, all exceptions are
now caught and raised as `WorkflowExecutionException` objects to
provide more detailed error information.