Fixed directory registration  (#881)
This commit fixes the incorrect registration of the directory in the
`DataManager`. Before this commit, the directory was not registered
because the `_register_path` method was called before the directory
was actually created (i.e., before the `await asyncio.gather()` call).
As a result, the check for the directory's existence in `_register_path`
caused the registration to fail.