Respect available CPUs in `LocalConnector` (#896)
Local execution: respect the number of available CPUs, not just
the # of installed CPUs. Note that `psutil.cpu_count()` may not
necessarily be equivalent to the actual number of CPUs the
current process can use. That can vary in case process CPU
affinity has been changed, Linux cgroups are being used or (in
case of Windows) on systems using processor groups or having
more than 64 CPUs. The number of usable CPUs can be obtained
with: `len(psutil.Process().cpu_affinity())`