Fix `StreamFlowContext` creation (#901)
This commit fixes the erroneous retrieval of the scheduler configuration
in the `StreamFlowContext` class and improves the related test.
It also adds schemas to the StreamFlow classes used for testing. This
operation involved a refactor of the tests because these schemas should
not be included in the default StreamFlow schema, otherwise they would
invalidate the checksum tests.