Fixed `ExpressionTool` outputs processing (#902)
This commit adds the `CWLExpressionOutputProcessor` to handle the
`ExpressionTool` outputs. Before this commit, StreamFlow was using a
`CWLCommandOutputProcessor` object, which accesses files on disk.
However, the CWL standard specifies that an `ExpressionTool` should not
have access to the file system. Additionally, this commit fixes the
`register_data` behavior for directories with a populated `listing` field.
Previously, these listings were skipped, resulting in missing registrations
of certain paths.